utilities/TempDecks_SetTemperature/requirements.txt
aiohttp==3.4.4
aionotify==0.2.0
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
future==0.18.2
idna==2.9
iso8601==0.1.13
jsonschema==3.2.0
multidict==4.7.6
multimethod==1.4
numpy==1.19.2
opentrons==3.21.0
opentrons-shared-data==3.21.0
pyrsistent==0.17.3
pyserial==3.4
PyYAML==5.3.1
six==1.15.0
typing-extensions==3.7.4.3
urwid==1.3.1
yarl==1.4.2
